pyproject.toml
[tool.poetry]
name = "sraandtools"
version = "0.1.0"
description = ""
authors = ["dar1593m <darwin.daume@hs-coburg.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.2.3"
pyserial = "^3.5"
numpy = "^2.2.2"
requests = "^2.32.3"
urllib3 = "^2.3.0"
plotly = "5.24.1"
dash = "^2.18.2"
scipy = "^1.15.1"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
apscheduler = "^3.11.0"
pydantic = "^2.10.6"
streamlit-plotly-events = "^0.0.6"
black = "^25.1.0"
openpyxl = "^3.1.5"
pdoc = "^15.0.1"
matplotlib = "^3.10.0"
altair = "^5.5.0"
scikit-learn = "^1.6.1"
seaborn = "^0.13.2"
tensorflow = "^2.15"
#tensorflow-io-gcs-filesystem = "0.27.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
